% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestest.R
\name{bfbinom}
\alias{bfbinom}
\title{Bayesian Binomial Test}
\usage{
bfbinom(data, p = 0.5, rscale = 0.1)
}
\arguments{
\item{data}{The vector containing binary data.}

\item{p}{Probability of one result.}

\item{rscale}{The r of the Prior Cauchy distribution.}
}
\description{
This function calculates Bayes Factors for every added data point of dichotomous data.
}
\details{
The first BF is calculated for 3 data points. For every subsequent data point a new BF is added.
The resulting BF vector indicates the change of evidence over time.
The function uses "proportionBF" from the BayesFactor package
}
\examples{
tbl$bf <- bfbinom(tbl@qbit)
}
