% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestest.R
\name{bfRobustness}
\alias{bfRobustness}
\title{Robustness Analysis for Bayesian Sequential tests}
\usage{
bfRobustness(x = NULL, informed = TRUE, prior.loc = NULL, prior.r = NULL)
}
\arguments{
\item{x}{A seqbf object created with bfttest(), bfbinom(), or bfcor()}

\item{informed}{Logical. If you don't want to specify prior locations and scales manually, informed = FALSE will test multiple prior widths and location 0, while informed = TRUE will test different locations and widths.}

\item{prior.loc}{Range of locations of the cauchy distributed prior function.}

\item{prior.r}{Range of scales of the cauchy distributed prior function.}
}
\description{
This function compares Bayes Factors of different priors. Returns a list containing BFMatrix, a data frame comprising the Bayes Factors of all possible combinations of the prior parameters (prior.location and prior.width)
}
\examples{
seqbf <- bfttest(rnorm(30, 0.5, 1), alternative = "two.sided", parallel = FALSE)

bfrInformed <- bfRobustness(seqbf) 
print(bfrInformed)

bfrUninformed <- bfRobustness(seqbf, informed = FALSE) 
plot(bfrInformed)
}
