% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestest.R
\name{bfttestRobustness}
\alias{bfttestRobustness}
\title{Robustness Analysis for Bayesian Sequential t-Test}
\usage{
bfttestRobustness(
  x = NULL,
  y = NULL,
  formula = NULL,
  data = NULL,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  prior.loc = seq(0.05, 1, 0.05),
  prior.r = seq(0.05, 1, 0.05)
)
}
\arguments{
\item{x}{A vector containing data values.}

\item{y}{A vector containing another set of values for a paired samples test.}

\item{formula}{A formula of the form var ~ group where var is a numeric variable giving the data values and group is indicating which group the data point belongs to (must be 2 levels). Use this for an independent samples t-Test.}

\item{data}{Use with formula. A data frame containing the variables given in the formula. Use this for an independent samples t-Test.}

\item{alternative}{Indicates the direction of the alternative hypothesis: two.sided, less, or greater}

\item{mu}{A number indicating the true value of the mean (or difference in means if you are performing a two sample test).}

\item{prior.loc}{Range of locations of the cauchy distributed prior function.}

\item{prior.r}{Range of scales of the cauchy distributed prior function.}
}
\description{
This function compares Bayes Factors of different priors. Returns a list containing BFMatrix, a data frame comprising the Bayes Factors of all possible combinations of the prior parameters (prior.location and prior.width)
Can be plotted with plotrobust().
}
\examples{
bfttestRobustness(sumscores, alternative = "greater") # One-sample
}
