% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeofevidence-package.R
\docType{package}
\name{changeofevidence-package}
\alias{changeofevidence}
\alias{changeofevidence-package}
\title{changeofevidence: Change of Evidence}
\description{
This package provides functions to test for a change of Bayesian evidence over time which can be used to examine volatile effects. It includes sequential Bayesian testing methods for t-Tests, Non-Parametric Tests, Binomial Tests, and Correlation Tests.
}
\keyword{internal}
