% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coetest.R
\name{fftcreate}
\alias{fftcreate}
\title{Create a Fast Fourier transformed vector}
\usage{
fftcreate(data)
}
\arguments{
\item{data}{A vector to be transformed.}
}
\description{
This function converts input signals into its Fast Fourier transform (FFT).
}
\details{
The FFT shows the spectral density of the input. It can be understood as a harmonic analysis and indicates the strength of osciallative components comprising the signal.
It can therefore be used to assess the meaning of oscillation as a characteristic. N/2 Frequencies are used as sampling rate.
}
\examples{
density.bf <- fftcreate(tbl$bf)
tblFFT <- data.frame(density.bf = fftcreate(tbl$bf), density.rw = fftcreate(tbl$rw))
}
