% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coetest.R
\name{fftlikelihood}
\alias{fftlikelihood}
\title{Frequency Analysis Likelihood Test}
\usage{
fftlikelihood(df, sims.df = sims, sims.df.col = "density.bf")
}
\arguments{
\item{df}{A dataframe containing the results of a ffttest.}

\item{sims.df}{A dataframe containing simulations, including columns "index" and "simid".}

\item{sims.df.col}{The column of the simulation dataframe that contains the comparison data.}
}
\value{
A vector containing the number of Top5-Frequencies of simulations.
}
\description{
This function estimates the likelihood of a specific proportion of Top5-Frequencies.
To serve as reference, the Top5-Frequencies of 10% of simulations in comparison to all simulations are calculated.
}
\examples{
fftlikelihood(r.fft)
}
