% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coetest.R
\name{maxbf}
\alias{maxbf}
\title{Maximum BF Analysis}
\usage{
maxbf(data, sims.df = sims)
}
\arguments{
\item{data}{A vector containing Bayes Factors.}

\item{sims.df}{A dataframe containing simulations, including columns "simid" and "bf".}
}
\value{
A list containing the Maximum BF of the experimental data, its position in the temporal order of data points, and the proportion of simulations with higher BFs at any time.
}
\description{
This function evaluates the likelihood of the highest reached Bayes Factor.
}
\details{
The highest reached (Maximum) BF of the experimental data is compared to those of all simulations.
An unusual peak can be assumed if less than 5\% of simulations show a higher BF at any time.
}
\examples{
r.maxbf <- maxbf(tbl$bf)
r.maxbf <- maxbf(tbl$bf, sims.df = newsims)
}
