% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coe_rw.R
\name{maxrw}
\alias{maxrw}
\title{Maximum Random Walk Analysis}
\usage{
maxrw(rw, sims.df = NULL)
}
\arguments{
\item{rw}{A numeric vector containing the random walk trajectory.}

\item{sims.df}{A dataframe containing simulations with columns "simid" and "rw".}
}
\value{
A list containing the maximum absolute RW value, its position, the signed
  value at that position, and the proportion of simulations with a higher peak.
}
\description{
Evaluates the likelihood of the highest absolute excursion reached by a random walk.
}
\details{
The peak absolute value of the observed random walk is compared to those of all
simulations. An unusual peak can be assumed if fewer than 5% of simulations show
a larger absolute excursion at any point.
}
\examples{
\dontrun{
r.maxrw <- maxrw(cumsum(rbinom(100, 1, 0.5) * 2 - 1), sims)
}
}
