% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotfft}
\alias{plotfft}
\title{Plot Fast Fourier Transform}
\usage{
plotfft(data, sims.df = NULL, sims.df.col = "density.bf", n.hz = 50,
  color = 3)
}
\arguments{
\item{data}{A vector containing Fourier transformed spectral densities.}

\item{sims.df}{A dataframe containing simulations, including column "simid". Set to NULL if you don't want to display simulations.}

\item{sims.df.col}{The name of the column of the simulation dataframe to compare to.}

\item{n.hz}{The amount of frequencies to be displayed.}
}
\description{
This function allows to plot a Fast Fourier Transform.
}
\details{
The FFT can be plotted by itself or in comparison to simulated data sets. GGplot2 is used to draw an image.
An experimental 95% Confidence Interval is calculated. Values above the black dotted line indicate Top5-amplitudes.
Use the full simulation set to correcly identify the 5% border line.
}
\examples{
p.fftbf <- plotfft(tblFFT$density.bf, color = 2)
p.fftbf

p.fftrw <- plotfft(tblFFT$density.rw, sims.df = sims, sims.df.col = "density.rw", color = 1)
p.fftrw
}
