% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotrobust}
\alias{plotrobust}
\title{Plot BF Robustness Analyses}
\usage{
plotrobust(data, color = "cornflowerblue")
}
\arguments{
\item{data}{An object generated with bfRobustness containing a "BFMatrix" dataframe.}
}
\description{
This function allows to plot a BF robustness analysis (bfttestRobustness())
}
\details{
Returns a heatmap showing the the BFs for different prior parameters (location and scale)
}
\examples{
plotrobust(bfRobustness(seqbf))
}
