% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotrobustTile}
\alias{plotrobustTile}
\title{Tile Plot BF Robustness Analyses}
\usage{
plotrobustTile(data, limit = 10)
}
\arguments{
\item{data}{An object generated with the bfRobustness()-function.}

\item{limit}{A BF limit which is marked in the plot.}
}
\description{
This function allows to plot a BF robustness analysis (bfttestRobustness())
}
\details{
Returns a heatmap showing the the BFs for different prior parameters (location and scale)
}
\examples{
\dontrun{
plotrobustTile(bfRobustness(seqbf))
}
}
