% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createsims.R
\name{print.coe_sim}
\alias{print.coe_sim}
\title{Print method for coe_sim objects}
\usage{
\method{print}{coe_sim}(x, ...)
}
\arguments{
\item{x}{A coe_sim object}

\item{...}{Additional arguments passed to print.data.frame}
}
\description{
Displays simulation metadata and first few rows of data
}
