% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createsims.R
\name{simcreate}
\alias{simcreate}
\title{Generate simulations for comparison}
\usage{
simcreate(
  trials,
  n.sims = 1000,
  mean.scores = NULL,
  method = c("pseudo", "files", "quantis"),
  filespath = "RandomFiles/",
  parallel = TRUE,
  nstart = 5,
  alternative = c("two.sided", "less", "greater"),
  prior.loc = 0,
  prior.r = 0.1,
  p = 0.5,
  use.files = NULL,
  use.quantis = NULL
)
}
\arguments{
\item{trials}{The amount of trials in a single experiment (this includes all subjects).}

\item{n.sims}{The amount of simulations to be generated. 10,000 is recommended, this might take several hours or days, though, depending on the sample size and number of trials.}

\item{mean.scores}{Should bits be summed up to represent a normal distribution centered about a mean? If yes, indicate the desired mean score here. If you are comparing to binomial data, set to NULL.}

\item{method}{The method used to generate the random data. Options are "pseudo" (software-based pseudo-RNG), "files" (text-files containing random bit sequences) and "quantis" (trueRNG device). Default is "pseudo".}

\item{filespath}{If random files should be used indicate the path to those files.}

\item{parallel}{If set to TRUE, multiple cores are being used in parallel to generate the simulations (recommenden).}

\item{nstart}{Number of data points that are considered before calculating the first BF (min = 2)}

\item{alternative}{Set parameter for Bayesian testing (t-Test).}

\item{prior.loc}{Set parameter for Bayesian testing (t-Test).}

\item{prior.r}{Set parameter for Bayesian testing.}

\item{p}{Set parameter for Bayesian testing (Binomial) or data generation for probabilities of success other than 0.5.}
}
\value{
A dataframe with trials*nsims rows containing the columns "simid","index","rw","density.rw","bf" and "density.bf".
}
\description{
This function generates a dataframe containing multiple simulations of random data.
}
\details{
The tests of this package rely on comparisons of the experimental data to simulated random data sets.
This function can provide a dataframe containing multiple simulated runs of the experiment with completely random data.
To generate the random data a software-based pseudo-RNG can be used or - better - the package provides text-files containing random bit sequences previously generated by a quantum-based trueRNG (Quantis).
Simulations can be created for binomial datasets and bits can be summed up to represent normally distributed mean scores.
}
\examples{
sims.pseudo <- simcreate(trials = 100, n.sims = 1000, method = "pseudo")
sims.files <- simcreate(trials = 100*20, n.sims = 1000, mean.scores = 10, method = "files", filespath = "RandomFiles/")
sims.quantis <- simcreate(trials = 1000, method = "quantis")
sims.parallel <- simcreate(trials = 376*100, n.sims = 1000, mean.scores = 50, method = "pseudo", parallel = TRUE) # 376 participants with 100 summed up trials each
}
