% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coetest.R
\name{simredo}
\alias{simredo}
\title{Create a copy of simulation-df with different amount of trials.}
\usage{
simredo(df, n, rw = TRUE)
}
\arguments{
\item{df}{A dataframe containing monte-carlo simulations. Must include the columns 'bf', 'density.bf', 'index', and 'simid'.}

\item{n}{The amount of trials per simulation. Must be smaller than in the original simulation df.}

\item{rw}{boolean. Set to TRUE if you also want to recalculate the density of the Random Walk. Needs the colums 'rw' and 'density.rw'.}
}
\description{
This function copies the BFs of an already computated simulation dataframe and re-calculates only the density via a FFT.
}
\details{
Computating BFs for monte-carlo-simulations can take a very long time. You can reuse an already calculated simulation-df and cut off the amount of trials per simulation. This function re-calculates only the densities of BF and RW via the Fast Fourier transformation.
}
\examples{
sim.new <- simredo(sims, length(bf.new), rw=F)
}
