% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestest.R
\name{bfbinom}
\alias{bfbinom}
\title{Bayesian Sequential Binomial Test}
\usage{
bfbinom(data, p = 0.5, prior.r = 0.1, nullInterval = NULL, nstart = 5)
}
\arguments{
\item{data}{A vector containing binary data.}

\item{p}{Probability of one result.}

\item{prior.r}{The r of the Prior Cauchy distribution.}

\item{nullInterval}{Optional vector of length 2 containing lower and upper bounds of an interval hypothesis to test, in probability units.}

\item{nstart}{How many data points should be considered before calculating the first BF (min = 2)}
}
\description{
This function calculates Bayes Factors for every added data point of dichotomous data.
}
\details{
The first BF is calculated for nstart data points. For every subsequent data point a new BF is added.
The resulting BF vector indicates the change of evidence over time.
The function uses "proportionBF" from the BayesFactor package
}
\examples{
tbl$bf <- bfbinom(tbl@qbit)
}
