% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestest.R
\name{bfbinom}
\alias{bfbinom}
\title{Bayesian Sequential Binomial Test}
\usage{
bfbinom(data, p = 0.5, rscale = 0.1, nstart = 3, inc = 1)
}
\arguments{
\item{data}{A vector containing binary data.}

\item{p}{Probability of one result.}

\item{rscale}{The r of the Prior Cauchy distribution.}

\item{nstart}{Number of data points that are considered before calculating the first BF (min = 2)}

\item{inc}{Number of new data points that are considered for each BF.}
}
\description{
This function calculates Bayes Factors for every added data point of dichotomous data.
}
\details{
The first BF is calculated for nstart data points. For every subsequent data point a new BF is added.
The resulting BF vector indicates the change of evidence over time.
The function uses "proportionBF" from the BayesFactor package
}
\examples{
tbl$bf <- bfbinom(tbl@qbit)
}
