% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestest.R
\name{bfcor}
\alias{bfcor}
\title{Bayesian Sequential Correlation Test}
\usage{
bfcor(x, y, nullInterval = 0, prior.r = 0.1, nstart = 3)
}
\arguments{
\item{x}{A vector containing continous data.}

\item{y}{A second vector containing continous data.}

\item{prior.r}{Prior distribution (scaled beta)}

\item{nstart}{How many data points should be considered before calculating the first BF (min = 2)}

\item{nullIntervall}{optional vector of length 2 containing lower and upper bounds of an interval hypothesis to test, in correlation units}
}
\description{
This function calculates Bayes Factors for the correlation of datasets.
}
\details{
The first BF is calculated for nstart data points. For every subsequent data point a new BF is added.
The resulting BF vector indicates the change of evidence over time.
The function uses "correlationBF" from the BayesFactor package
}
\examples{
bfcor(exp$sums, con$sums, nullInterval = c(-1,0))
}
