% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestest.R
\name{bfRobustness}
\alias{bfRobustness}
\title{Robustness Analysis for Bayesian Sequential tests}
\usage{
bfRobustness(
  x = NULL,
  prior.loc = seq(0.05, 1, 0.05),
  prior.r = seq(0.05, 1, 0.05)
)
}
\arguments{
\item{x}{A seqbf object created with bfttest(), bfbinom(), or bfcor()}

\item{prior.loc}{Range of locations of the cauchy distributed prior function.}

\item{prior.r}{Range of scales of the cauchy distributed prior function.}
}
\description{
This function compares Bayes Factors of different priors. Returns a list containing BFMatrix, a data frame comprising the Bayes Factors of all possible combinations of the prior parameters (prior.location and prior.width)
}
\examples{
bfr <- bfRobustness(seqbf)
print(bfr)
plot(bfr)
}
