% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestest.R
\name{bfttest}
\alias{bfttest}
\title{Bayesian Sequential t-Test}
\usage{
bfttest(
  x = NULL,
  y = NULL,
  formula = NULL,
  data = NULL,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  prior.loc = 0,
  prior.r = 0.1,
  nstart = 5
)
}
\arguments{
\item{x}{A vector containing data values or a formula.}

\item{y}{A vector containing another set of values for a paired samples test.}

\item{formula}{A formula of the form var ~ group where var is a numeric variable giving the data values and group is indicating which group the data point belongs to (must be 2 levels). Use this for an independent samples t-Test.}

\item{data}{Use with formula. A data frame containing the variables given in the formula. Use this for an independent samples t-Test.}

\item{alternative}{Indicates the direction of the alternative hypothesis: two.sided, less, or greater}

\item{mu}{A number indicating the true value of the mean (or difference in means if you are performing a two sample test).}

\item{prior.loc}{Location of the cauchy distributed prior function (use 0 for an uninformed prior).}

\item{prior.r}{Scale of the cauchy distributed prior function.}

\item{nstart}{How many data points should be considered before calculating the first BF (min = 2).}
}
\description{
This function calculates Bayes Factors for every added data point of normally distributed data, e.g. sum scores.
}
\details{
The first BF is calculated for Nstart data points. For every subsequent data point a new BF is added.
The resulting BF vector indicates the change of evidence over time.
The function first calculates t-scores and p-values and subsequently uses the BFDA package to translate the data into BFs.
Returns a list of t-scores ($t-value), p-scores ($p-value), and BF ($BF)
}
\examples{
bfttest(sumscores, alternative = "greater") # One-sample

tbl$bf <- bfttest(tbl$sums, alternative = "two.sided", mu = 50, nstart = 10, prior.loc = 0.1, prior.r = 0.05)$BF

bfttest(formula = raw~group, data = df) # Independent samples

bfttest(df1$scores, df2$scores) # Paired samples
}
