% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestest.R
\name{bfttest}
\alias{bfttest}
\title{Bayesian Sequential t-Test}
\usage{
bfttest(
  data,
  ydata = NULL,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  paired = FALSE,
  prior.loc = 0,
  prior.r = 0.1,
  nstart = 5
)
}
\arguments{
\item{data}{A vector containing data values.}

\item{ydata}{A vector containing another set of values for a paired samples test (independent samples are currently not supported). Please indicate with paired = TRUE.}

\item{alternative}{Indicates the direction of the alternative hypothesis: two.sided, less, or greater}

\item{mu}{A number indicating the true value of the mean (or difference in means if you are performing a two sample test).}

\item{paired}{Logical. Set TRUE if you want to perform a paired samples t-Test.}

\item{prior.loc}{Location of the cauchy distributed prior function (use 0 for an uninformed prior).}

\item{prior.r}{Scale of thr cauchy distributed prior function.}

\item{nstart}{How many data points should be considered before calculating the first BF (min = 2).}
}
\description{
This function calculates Bayes Factors for every added data point of normally distributed data, e.g. sum scores.
}
\details{
The first BF is calculated for Nstart data points. For every subsequent data point a new BF is added.
The resulting BF vector indicates the change of evidence over time.
The function first calculates t-scores and p-values and subsequently uses the BFDA package to translate the data into BFs.
Returns a list of t-scores ($t-value), p-scores ($p-value), and BF ($BF)
}
\examples{
bflist <- bfttest(sumscores, alternative = "greater")

tbl$bf <- bfttest(tbl$sums, alternative = "two.sided", mu = 50, nstart = 10, prior.loc = 0.1, prior.r = 0.05)$BF
}
