% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coetest.R
\name{ffttest}
\alias{ffttest}
\title{Frequency Analysis Test}
\usage{
ffttest(data, sims.df = sims, sims.df.col = "density.bf", top5 = FALSE)
}
\arguments{
\item{data}{A vector containing Fourier transformed (spectral density) data (use 'fftcreate' function).}

\item{sims.df}{A dataframe containing simulations, including columns "index" and "simid".}

\item{sims.df.col}{The column of the simulation dataframe that contains the comparison data.}

\item{top5}{Logical. If set to TRUE, function will additionally return the Top5-Frequency method. For each frequency th function counts how many simulations show a higher amplitude. If no more than 5% of simulations are above the experimental value, it is considered a "Top5-Frequency". The proportion of Top5-Frequencies indicates the pronouncedness of oscillatory elements in the data.}
}
\value{
A list containing a dataframe of all frequencies and the proportion of simulations with a lower amplitude, and information on the sum of amplitudes.
}
\description{
This function analyses FFTs and compares the amplitudes of all frequencies to those of simulations.
}
\details{
If you want to use the old "Top5-Frequency" method instead of amplitude sums, indicate by setting 'top5 = TRUE'.
}
\examples{
r.fftbf <- ffttest(tblFFT$density.bf)
r.fftrw <- ffttest(tblFFT$density.rw, sims.df = newsims, sims.df.col = "density.rw")
}
