% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createsims.R
\name{simcreate}
\alias{simcreate}
\title{Generate simulations for comparison}
\usage{
simcreate(trials, n.sims = 10000, mean.scores = NULL,
  use.files = TRUE, filespath = "simfiles/", parallel = TRUE)
}
\arguments{
\item{trials}{The amount of trials in a single experiment (this includes all subjects).}

\item{n.sims}{The amount of simulations to be generated. 10,000 is recommended, this might take several hours or days, though, depending on the sample size and number of trials.}

\item{mean.scores}{Should bits be summed up to represent a normal distribution centered about a mean? If yes, indicate the desired mean score here. If you are comparing to binomial data, set to NULL.}

\item{use.files}{Should the random data be taken out of the tRNG files (TRUE) or should a pseudo RNG be used (FALSE).}

\item{filespath}{If random files should be used indicate the path to those files.}

\item{parallel}{If set to TRUE, multiple cores are being used in parallel to generate the simulations (recommenden).}
}
\value{
A dataframe with trials*nsims rows containing the columns "simid","index","rw","density.rw","bf" and "density.bf".
}
\description{
This function generates a dataframe containing multiple simulations of random data.
}
\details{
The tests of this package rely on comparisons of the experimental data to simulated random data sets.
This function can provide a dataframe containing multiple simulated runs of the experiment with completely random data.
To generate the random data a software-based pseudo-RNG can be used or - better - the package provides text-files containing random bit sequences previously generated by a quantum-based trueRNG (Quantis).
Simulations can be created for binomial datasets and bits can be summed up to represent normally distributed mean scores.
}
\examples{
sims <- simcreate(100)
sims1000 <- simcreate(100, n.sims = 1000)
sims <- simcreate(600, mean.scores = 30)
sims.pseudo <- simcreate(2000, use.files = FALSE)
}
